+++ plugins/integrations/confluence/plugin_test.go
@@ -3,6 +3,7 @@ package confluence
 import (
 	"context"
 	"encoding/base64"
+	browser "github.com/fluxplane/fluxplane-browser"
 	"strings"
 	"testing"
 
@@ -296,7 +297,7 @@ type fakeSystem struct {
 func (s fakeSystem) Workspace() system.Workspace     { return nil }
 func (s fakeSystem) Network() system.Network         { return s.network }
 func (s fakeSystem) Process() system.ProcessManager  { return nil }
-func (s fakeSystem) Browser() system.BrowserManager  { return nil }
+func (s fakeSystem) Browser() browser.Manager        { return nil }
 func (s fakeSystem) Clarifier() system.Clarifier     { return nil }
 func (s fakeSystem) Environment() system.Environment { return s.env }