+++ plugins/integrations/gitlab/plugin_test.go
@@ -3,6 +3,7 @@ package gitlab
 import (
 	"context"
 	"encoding/base64"
+	browser "github.com/fluxplane/fluxplane-browser"
 	"regexp"
 	"strconv"
 	"strings"
@@ -3735,7 +3736,7 @@ func (s fakeSystem) Network() system.Network { return s.network }
 
 func (s fakeSystem) Process() system.ProcessManager { return nil }
 
-func (s fakeSystem) Browser() system.BrowserManager { return nil }
+func (s fakeSystem) Browser() browser.Manager { return nil }
 
 func (s fakeSystem) Clarifier() system.Clarifier { return nil }