+++ plugins/integrations/web/search_test.go
@@ -2,6 +2,7 @@ package web
 
 import (
 	"context"
+	browser "github.com/fluxplane/fluxplane-browser"
 	"sync"
 	"testing"
 
@@ -19,7 +20,7 @@ type testSystem struct {
 func (s testSystem) Workspace() system.Workspace     { return s.workspace }
 func (s testSystem) Network() system.Network         { return s.network }
 func (s testSystem) Process() system.ProcessManager  { return nil }
-func (s testSystem) Browser() system.BrowserManager  { return nil }
+func (s testSystem) Browser() browser.Manager        { return nil }
 func (s testSystem) Clarifier() system.Clarifier     { return nil }
 func (s testSystem) Environment() system.Environment { return s.env }