+++ runtime/language/toolchain_test.go
@@ -2,6 +2,7 @@ package language
 
 import (
 	"context"
+	browser "github.com/fluxplane/fluxplane-browser"
 	"testing"
 	"time"
 
@@ -56,7 +57,7 @@ type fakeSystem struct {
 func (s fakeSystem) Workspace() system.Workspace     { return nil }
 func (s fakeSystem) Network() system.Network         { return nil }
 func (s fakeSystem) Process() system.ProcessManager  { return s.process }
-func (s fakeSystem) Browser() system.BrowserManager  { return nil }
+func (s fakeSystem) Browser() browser.Manager        { return nil }
 func (s fakeSystem) Clarifier() system.Clarifier     { return nil }
 func (s fakeSystem) Environment() system.Environment { return nil }